Cap 5 data embeds per message
Discord technically has a limit of 10 embeds per message, but there's also a limit on the total combined size of all embeds. Because the bot's embeds tend to be quite large with lots of info, I decided to cap the number of embeds it would send in one message to not run up against Discord's embed length limit (and also prevent spam).