Fix Neuron art resolution
For cards that have released on the database but aren't on Master Duel, the bot uses the YGOrg artwork repository to get their Neuron artwork and cache it by saving it to disk.

Unfortunately the lookup for cached images on disk wasn't working due to an incorrect absolute path being provided. Also, to make matters worse, the artwork repo manifest is supposed to auto-evict after 24 hrs and then be re-cached when necessary, but that re-cache was not working because the auto-evict wasn't evicting properly.