Rework YGOrg DB caching
Previously the manifest was checked on every request and processed if a new one was found. Now the manifest will simply be checked on a periodic and updated accordingly. Updates to the manifest and evictions from the cache will ONLY be done through this periodic to avoid cache invalidation conflicts.

Also, on cache eviction from manifest updates, the bot will attempt to re-cache the (new) value provided by the API. Previously it would evict and only re-cache if the data was requested again. Put simply, the bot will be more proactive when caching, instead of only caching on request.