axios -> fetch, fix artwork repo queries
Moved all mentions of axios to fetch (with associated refactoring) in all files except TCGPlayerHandler, which needs a lot more work. This move was done because there is a memory leak somewhere and quick googling implies axios has been found to have some memory leaks.

Also fixed some issues with artwork repo queries (sometimes it was querying when it didn't need to) and especially the associated save of the raw data from the repo.