Decrease discord.js caching
The bot has had a long-standing issue with indefinitely increasing memory usage. On the AWS server the official instance runs on, where memory is relatively limited, this eventually causes the bot to crash.

I thought this was a memory leak for a long time but attempts to troubleshoot have been ineffective. Recent research pointed me in a different direction and that turned up some information about discord.js's aggressive persistent caching of guilds, users, channels, messages, etc., as well as the configuration it offers to decrease its cache size and lifetimes for certain data types, like limiting the number of messages per channel or members per server that can be cached at any one time.

Local testing with new caching limitations indicates a significantly reduced memory footprint (at least in terms of the heap) over long periods of time. RSS still seems to increase indefinitely, but currently I believe that's as much as Node problem as a bot problem. Memory on my local test bench is much more freely available, so this commit is to officially enable the limited caching to see if it resolves OOM issues on the AWS server.