YGOrg -> YGOResources migration + logging cleanup
The YGOrg DB is now named YGOResources DB, with associated URL changes to their API and website. This commit removes all code references to the old YGOrg DB in favor of YGOResources, with the one exception of the link regexes so that links to the old YGOrg DB are still seen by the bot and treated as queries to YGOResources.

Also used the opportunity to do some logging and import cleanup, it was long overdue.