Price check and YGOResources ruling updates
Found an edge case where the TCGPlayer name for a card was different from its database name, causing it to not turn up any price results. Implemented an alias map, unfortunately hardcoded, to be expanded if more examples are found.

Also, YGOResources added an indicator within rulings when a massive list of card names is about to begin. Updated the bot logic to check for that separator and prune the massive list to make some ruling embeds much smaller.