Fix broken FAQ blocks
The old logic for breaking up FAQ blocks for display in a discord embed was faulty and missed an edge case that caused the first block to be too long for the 1024 character field limit. That's been fixed and the FAQ embed generation in general has been improved to be more robust against stuff like that.