Price rarity filter update + ignore markdown links
Markdown links like [this example](https://link-to-website.com) are now usable for regular Discord users (previously they were restricted to bots). Naturally, the leading [text in brackets] conflicts with the bot's default syntax, so I've added this markdown syntax to the list of things the bot will ignore and not interpret as a query. For servers that don't even use the default open/close bracket syntax, this change won't affect anything.

Also, it was brought to my attention that the price filters in the /price command were sometimes too general. For example, trying to filter on "Secret Rare" was returning both Secret Rares and any other rarity with "Secret Rare" as a substring, like Gold Secret Rare or Quarter Century Secret Rare. This warranted an update to the filtering functionality to correctly distinguish between literal rarity searches and "substring" ones of which multiple variants exist.