Fix "Unreleased" designation on some cards
For cards with a "latest print" that was in the future, i.e., an upcoming print that wasn't yet officially printed, the bot logic was mistakenly flagging them as "unreleased" in the footer that lists the card's status in all CGs. This was due to some unnecessarily strict logic that was flagging ANY future print as being the mark of an unreleased card, despite the possibility that it had previous prints and it was just the latest that was yet to be printed.

Fixed this mistaken logic by flipping the check around; now, instead of looking for whether the latest print is in the future and (incorrectly) flagging the card as unreleased if it was, the bot instead looks for ANY print in the past, and only flags it as unreleased if no past print exists.