Fix Yugipedia queries + links
A recent change to the API returns of Yugipedia was causing issues with card embeds. They changed how effect text was referred to in the raw JSON from "lore" to "text", so the bot needed to be updated to look for this new field.

Also, fixed an edge case issue recently introduced with the Maliss cards that was causing links to Yugipedia to be broken due to invalid URI characters.