Change card attribute + property display
Before this commit, monster attribute and spell/trap property was displayed in an icon next to the embed name. This was reported to cause problems for people with disabilities or colorblindness, as due to their small size it could be difficult to actually tell what the icon was. Also, in some cases like for Continuous Spell/Traps, this created an ambiguous display where it was difficult to determine whether the card was a Spell or Trap, since either could be Continuous.

The resolution for this has been to completely remove the icon from embeds, and instead identify the card's attribute/properties in the description below. Icons are still used via discord emotes, but they are included alongside text that explicitly reports the attribute or properties of the card, e.g., Continuous Spell or Counter Trap. All the additional text is also properly translated to other locales as necessary.