Fix Yugipedia parsing edge cases
Some old (especially weird) cards, mostly from old games, have outdated structures in the API and store standard data (like effect text) in fields the bot doesn't look for, which breaks things downstream and causes the bot to error out when generating data embeds to respond with.

This commit adds special cases to check for these edge case fields in case they're needed. There may be more that need to be added in the future.