Fix function call on null object
The recent change to Spell/Trap and monster attribute display required a toLowerCase call on the card type, but this call was done without first checking whether the type is null.

For well-formed cards this isn't an issue, but Yugipedia queries don't guarantee well-formed cards so this was causing issues when bot queries resolved to some Yugipedia data.