Support for Master Duel alt arts
I haven't paid attention to Master Duel since adding support for its high res artwork. Imagine my surprise when I find out that some MD cards have alt arts, meaning they have multiple high res artworks!

Adding support for it was fairly easy in code, but the process of actually exporting and FINDING the alt artworks to be stored locally was a massive pain in the ass. That all happened in the background of this commit, so supporting this was not as easy as this small commit might make it seem.

For future reference: (almost) all Master Duel base artworks are identified by the card's Konami database ID, which makes matching them with their card very easy. However, alt artworks have (as far as I can tell) random IDs, so there's no obvious way to match them to their card, nor any metadata within the images that could be used to match by card name. So the process of assocating alt arts with their cards involved a whole lot of manual work, including image comparisons with the 10k+ artworks that currently exist. It was actually a fun problem to solve but it requires manual work to maintain into the future, so I hope MD doesn't add alt arts so often that it will drive me crazy.