Added art command options for source + art
The art command previously has only allowed querying for a particular card, and responded with an interactable series of selectors allowing the selection of source and art ID. All of this interactivity was hidden in an ephemeral (i.e., only viisble to the command user) message, and the final product was only shared with chat when the user clicked a "Post to Chat" button.

This was an extremely clunky compromise to solve an issue that was created when the original interaction message was not ephemeral. In that case, the chat would move up and down as the user tinkered with which art they wanted to view, which could be very annoying to other users in the chat.

With this update, "source" and "art" can now be specified in the original command, and if both are given (or a specific art + source combo is chosen with "art"), then the interaction is skipped and the result is posted straight to chat.